code restructuring done
