--- src/components/Analytics/data.js
@@ -1,42 +1,0 @@
-export const thArray = ['Connected Website', 'Total Visitors', 'User Signups', 'Total Live Stats', 'Conversion %', 'Action'];
-export const tdArray = [
-  { 'website': 'http://www.socialtrooper.com', 'tvisitor': '313', 'signups': '30', 'liveuser': '5' },
-  { 'website': 'http://www.shahidkhan.in', 'tvisitor': '113', 'signups': '30', 'liveuser': '5' },
-  { 'website': 'http://www.socialtrooper.com', 'tvisitor': '213', 'signups': '10', 'liveuser': '5' },
-  { 'website': 'http://www.shahidkhan.in', 'tvisitor': '23', 'signups': '20', 'liveuser': '5' },
-  { 'website': 'http://www.socialtrooper.com', 'tvisitor': '223', 'signups': '40', 'liveuser': '5' },
-];
-
-export const profileHeader = [
-  'Photo', 'Name', 'Email', 'Location', 'Country', 'Sign up on', 'Action'
-];
-
-export const profileData = [
-  { 'image': 'http://via.placeholder.com/140X100', 'Name': 'Notifiaction1', 'Email': 'test@test.com', 'Location': 'Delhi', 'country': 'India', 'social': 'yes', 'pview': '200', 'lastseen': '03/03/2018 09:51AM', 'firstseen': '01/03/2018 07:51AM' },
-  { 'image': 'http://via.placeholder.com/140X100', 'Name': 'Notifiaction1', 'Email': 'test@test.com', 'Location': 'Delhi', 'country': 'India', 'social': 'yes', 'pview': '200', 'lastseen': '03/03/2018 09:51AM', 'firstseen': '01/03/2018 07:51AM' },
-  { 'image': 'http://via.placeholder.com/140X100', 'Name': 'Notifiaction1', 'Email': 'test@test.com', 'Location': 'Delhi', 'country': 'India', 'social': 'yes', 'pview': '200', 'lastseen': '03/03/2018 09:51AM', 'firstseen': '01/03/2018 07:51AM' },
-  { 'image': 'http://via.placeholder.com/140X100', 'Name': 'Notifiaction1', 'Email': 'test@test.com', 'Location': 'Delhi', 'country': 'India', 'social': 'yes', 'pview': '200', 'lastseen': '03/03/2018 09:51AM', 'firstseen': '01/03/2018 07:51AM' }
-];
-
-export const chartdata = [
-  { name: '10/03/2018', click: 4000, visit: 2400, active: 10 },
-  { name: '11/03/2018', click: 3000, visit: 1398, active: 1200 },
-  { name: '12/03/2018', click: 2000, visit: 9800, active: 1250 },
-  { name: '13/03/2018', click: 2780, visit: 3908, active: 6200 },
-  { name: '14/03/2018', click: 1890, visit: 4800, active: 1140 },
-  { name: '15/03/2018', click: 2390, visit: 3800, active: 4120 },
-  { name: '16/03/2018', click: 3490, visit: 4300, active: 7145 },
-];
-
-export const journeyHeader = [
-  'Date', 'Time of Visit', 'Page Visited', 'Link / Button Clicked', 'Time Spent on Page'
-];
-export const journeyData = [
-  { 'date': '03/03/2018', 'time': '03:00 PM', 'visit': '150', 'click': '150', 'hour': '150 Min' },
-  { 'date': '03/03/2018', 'time': '03:00 PM', 'visit': '150', 'click': '150', 'hour': '150 Min' },
-  { 'date': '03/03/2018', 'time': '03:00 PM', 'visit': '150', 'click': '150', 'hour': '150 Min' },
-  { 'date': '03/03/2018', 'time': '03:00 PM', 'visit': '150', 'click': '150', 'hour': '150 Min' },
-  { 'date': '03/03/2018', 'time': '03:00 PM', 'visit': '150', 'click': '150', 'hour': '150 Min' }
-
-
-];
