+++ src/components/Dashboard/Website.js
@@ -5,7 +5,6 @@ export class Website extends Component{
 
     handleCheckbox = event => {
       const target = event.target;
-      console.log(event.target);
       this.setState({
         [target.name]: target.checked
       });
