+++ src/components/NotificationConfigure/Channel/Channel.js
@@ -36,7 +36,6 @@ class Channel extends Component  {
     });
   }
 
-
   render() {
     return (
       <div  className="channel-container">
@@ -44,7 +43,6 @@ class Channel extends Component  {
         <Col md={12}>
           {this.channelsList()}
         </Col>
-
       </div>
     );
   }
