+++ src/components/NotificationConfigure/Channel/Channel.scss
