+++ src/components/NotificationConfigure/NotificationConfigure.scss
