+++ src/components/NotificationConfigure/NotificationDesign/NotificationDesign.scss
