+++ src/components/NotificationConfigure/NotificationDesignSetting/NotificationDesignSetting.scss
