+++ src/components/NotificationSettings/NotificationSettings.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { Row } from 'react-bootstrap';
-import NotificationConfigure from './NotificationConfigure/NotificationConfigure';
+import NotificationConfigure from 'components/NotificationConfigure/NotificationConfigure';
 import NotificationList from './NotificationList/NotificationList';
 
 import './NotificationSettings.scss';
@@ -225,7 +225,7 @@ class Notifications extends Component {
       <div className="notification-settings">
         <div>
           <h4 className="lead text-center m-b-30 m-t-20">Notifications</h4>
-        
+
 
           {!this.props.notification
             ?
