+++ src/components/ReviewRedirect/ReviewRedirect.js
@@ -8,17 +8,12 @@ import './ReviewRedirect.scss';
 
 class ReviewRedirect extends React.Component {
   componentDidMount() {
-    // console.log(this.props.campaign.trackingId);
-
     const {
       params: { provider},
       location: { search},
     } = this.props;
     const requestURL = `integrations/${provider}/callback${search}`;
-    console.log(provider,'Provider');
-    console.log(search,'Location');
     this.props.reviewRedirect(requestURL);
-    console.log(requestURL,'requestURL');
   }
 
   render() {
