+++ src/components/WebsiteDemoPage/WebsiteDemoPage.js
@@ -65,7 +65,6 @@ class WebsiteDemoPage extends  Component  {
   }
 
   handleStateChange = (target, value) => {
-    console.log(target, value);
     this.setState({[target]: value});
   }
