+++ src/components/index.js
@@ -47,6 +47,7 @@ export { default as Rules } from './Rules/Rules';
 export { default as DisplayPage } from './DisplayPage/DisplayPage';
 export { default as CapturePage } from './CapturePage/CapturePage';
 export { default as InstallPixel } from './InstallPixel/InstallPixel';
+export { default as NotificationConfigure } from './NotificationConfigure/NotificationConfigure';
 
 export { default as Oauthpage } from './Oauthpage/Oauthpage';
 export { default as Oauthgenerate } from './Oauthgenerate/Oauthgenerate';
